RecycleView not working yet, MainActivity working

+++ app/src/main/java/com/example/watertrackerapp/WaterIntake.kt
@@ -1,6 +1,6 @@
 package com.example.watertrackerapp
 
 data class WaterIntake(
-    val amount: Double,  // ilość wypitej wody w ml
+    val amount: Int,  // ilość wypitej wody w ml
     val date: String     // data w formacie "yyyy-MM-dd"
 )
